+++ c/apps/temperature.h
@@ -2,6 +2,6 @@
 #define TEMPERATURE_H_INCLUDED_ 1
 
 int temperature_celcius_to_fahrenheit(int celcius);
-int temperature_fahrenheit_to_celcius(int fahrenheit);
+float temperature_fahrenheit_to_celcius(float fahrenheit);
 
 #endif  /* !defined(TEMPERATURE_H_INCLUDED_) */
